GUI: Currency with 2 Digits

+++ src/main/java/ch/hemisoft/immo/domain/PurchaseCost.java
@@ -1,5 +1,7 @@
 package ch.hemisoft.immo.domain;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Embeddable;
 import javax.validation.constraints.Min;
 
@@ -8,13 +10,13 @@ import lombok.Data;
 @Data
 @Embeddable
 public class PurchaseCost {
-	@Min(0)		double landAcquisition;
-	@Min(0)		double notary;
-	@Min(0)		double agency;
-	@Min(0)		double valuation;
-	@Min(0)		double court;
+	@Min(0)		BigDecimal landAcquisition;
+	@Min(0)		BigDecimal notary;
+	@Min(0)		BigDecimal agency;
+	@Min(0)		BigDecimal valuation;
+	@Min(0)		BigDecimal court;
 	
 	public double getTotalCompletionCost() {
-		return landAcquisition + notary + agency + valuation + court;
+		return landAcquisition.doubleValue() + notary.doubleValue() + agency.doubleValue() + valuation.doubleValue() + court.doubleValue();
 	}
 }
