+++ src/main/java/ch/hemisoft/immo/domain/RunningCost.java
@@ -1,5 +1,7 @@
 package ch.hemisoft.immo.domain;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Embeddable;
 import javax.validation.constraints.Min;
 
@@ -8,6 +10,6 @@ import lombok.Data;
 @Data
 @Embeddable
 public class RunningCost {
-	@Min(0)		double administrationEachApartment;
-	@Min(0)		double maintenanceEachQm;
+	@Min(0)		BigDecimal administrationEachApartment;
+	@Min(0)		BigDecimal maintenanceEachQm;
 }
